Discriminant Problem - Segundo
