+++ lsql/judge/models.py
@@ -21,10 +21,10 @@ from django.core.serializers.json import DjangoJSONEncoder
 
 from .feedback import compare_select_results, compare_db_results, compare_function_results, compare_discriminant_db
 from .oracle_driver import OracleExecutor
-from .types import VeredictCode, ProblemType
+from .types import VeredictCode, ProblemType, OracleStatusCode
 from .parse import load_select_problem, load_dml_problem, load_function_problem, load_proc_problem, \
     load_trigger_problem
-from .exceptions import ZipFileParsingException
+from .exceptions import ZipFileParsingException, ExecutorException
 
 
 def markdown_to_html(markdown_text, remove_initial_p=False):
@@ -418,7 +418,7 @@ class TriggerProblem(Problem):
 
 
 class DiscriminantProblem(Problem):
-    """Problem that requires an INSERT as solution"""
+    """Problem that requires an INSERT as solution for debug an incorrect query"""
     check_order = models.BooleanField(default=False)
     correct_query = models.TextField(max_length=5000, validators=[MinLengthValidator(1)], blank=True)
     incorrect_query = models.TextField(max_length=5000, validators=[MinLengthValidator(1)], blank=True)
@@ -429,25 +429,31 @@ class DiscriminantProblem(Problem):
 
     def clean(self):
         """Executes the problem and stores the expected result"""
-        super().clean()
-        executor = OracleExecutor.get()
-        self.expected_result = []
-        self.initial_db = []
-        for insert_sql in self.insert_sql_list():
-            res = executor.execute_select_test(self.create_sql, insert_sql,
-                                               self.incorrect_query, output_db=True)
-            self.expected_result.append(res['result'])
-            self.initial_db.append(res['db'])
+        try:
+            super().clean()
+            executor = OracleExecutor.get()
+            self.expected_result = []
+            self.initial_db = []
+            # In this case (this type of problem) there are only one database
+            for insert_sql in self.insert_sql_list():
+                res = executor.execute_select_test(self.create_sql, insert_sql,
+                                                   self.incorrect_query, output_db=True)
+                self.expected_result.append(res['result'])
+                self.initial_db.append(res['db'])
+        except Exception as excp:
+            raise ValidationError(excp) from excp
 
     def judge(self, code, executor):
-        i_query = executor.execute_select_test(self.create_sql, self.insert_sql + code,
-                                               self.incorrect_query, output_db=False)
-        c_query = executor.execute_select_test(self.create_sql, self.insert_sql + code,
-                                               self.correct_query, output_db=False)
-        veredict, feedback = compare_discriminant_db(i_query['result'], c_query['result'], self.check_order)
-        if veredict != VeredictCode.AC:
-            return veredict, feedback
-        return VeredictCode.AC, ''
+        try:
+            # In this case (this type of problem) there are only one database
+            for insert_sql in self.insert_sql_list():
+                i_query = executor.execute_select_test(self.create_sql, insert_sql + code,
+                                                       self.incorrect_query, output_db=False)
+                c_query = executor.execute_select_test(self.create_sql, insert_sql + code,
+                                                       self.correct_query, output_db=False)
+            return compare_discriminant_db(i_query['result'], c_query['result'], self.check_order)
+        except ExecutorException as excp:
+            raise ExecutorException(OracleStatusCode.EXECUTE_USER_CODE) from excp
 
     def problem_type(self):
         return ProblemType.DISC
